Изменения получения координат

+++ location_beacons_android/android/src/main/kotlin/ru/elocont/location_beacons/location_beacons/models/response/CellLocation.kt
@@ -5,7 +5,7 @@ import com.squareup.moshi.Json
 data class CellLocation(
     val status: String,
     val message: String?,
-    val accuracy: Int?,
+    val accuracy: Double?,
     val address: String?,
 
     @Json(name = "lat")
